# PERF Grey dilation