As discussed, here a different `local_maxima` that does a grey dilation with a rectangular kernel, instead of a square one. Directly after, it eliminates duplicates using a KDTree.

I also refactored the code a bit in view of #257  : the `find_maxima` is renamed and moved to `find.py`.