@nkeim 2016-04-29
Excellent.

Notes on the ongoing performance comparison with MATLAB, for sparsely-populated frames:

- Number of features found went from 546 to 554. The 2-norm metric used by KDtree ≠ metric in pixellated space. But a ~1% difference is probably acceptable.
- Feature-finding time went from 0.106 s/frame to 0.0517 s/frame (!). MATLAB's current best time is 0.0458 s. A full 10 ms of trackpy's time is still used by pims' `as_grey`.
- The final product *looks* worse with the same parameters. For example, I draw your attention to the lower right corner of each image:

Before

![benchmarking_sparse_trackpy_after_e20c6c4cf2f46966d92a04e2f50acb62975049ff](https://cloud.githubusercontent.com/assets/2985325/14930512/43c67a94-0e19-11e6-8311-b70f122201ff.png)

After

![sparse_trackpy](https://cloud.githubusercontent.com/assets/2985325/14930523/52e27c58-0e19-11e6-9910-75ae146ff2d8.png)

So there are some issues to address, but this is a very successful experiment!