@nkeim 2016-04-29
Bad news: this slows down the dense benchmark by about 25x. Profiler output for 10 frames:

```
         23297539 function calls (23297159 primitive calls) in 85.399 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       20   50.948    2.547   80.919    4.046 find.py:18(where_close)
       20   16.450    0.823   16.450    0.823 {method 'query_pairs' of 'ckdtree.cKDTree' objects}
 11638432    4.137    0.000    4.137    0.000 find.py:34(<genexpr>)
 11638432    4.068    0.000    4.068    0.000 find.py:33(<genexpr>)
       40    4.012    0.100   12.217    0.305 {numpy.core.multiarray.fromiter}
       10    2.247    0.225   82.594    8.259 find.py:53(drop_close)
       20    0.818    0.041    0.818    0.041 {method 'sort' of 'numpy.ndarray' objects}
       10    0.397    0.040   83.484    8.348 find.py:70(grey_dilation)
       58    0.346    0.006    0.346    0.006 {numpy.core.multiarray.where}
       20    0.324    0.016    0.324    0.016 {scipy.ndimage._nd_image.correlate1d}
       20    0.287    0.014    0.287    0.014 {scipy.ndimage._nd_image.uniform_filter1d}
```

Given the badness of this worst-case scenario, I think we should find a way to make this new method optional.