@caspervdw 2016-04-30
Sorry about the unintended close.
That's the reason why I changed `where_close`. The iterative algorithm was even slower than this. I convinced myself that iterations are not necessary but tests are certainly a good idea.

Is the benchmark a real image or a random noise?