@caspervdw 2016-04-30
@nkeim I discovered a bug that might be crucial for the benchmark. I can't manage to put in a line note on my phone. The call to `grey_dilation` in `locate` should have `diameter` as second arg.

You coulf even argue to put in `separation` as second arg, but that's a big change.