@nkeim 2016-05-01
Ah! With that bugfix made, number of features now goes from 546 (old method) to 351 (corrected new method). So maybe there is an off-by-one error somewhere.

The dense benchmark is still 10x slower than before. This is an image of close-packed repulsive particles, with typical spacing just a little bigger than feature size. So it is a worst case for this algorithm, but one I would argue is not uncommon.

I am (slowly) looking at alternatives to cKDTree for this. The precision and isotropy it offers are not really necessary at this stage. I like the Blair-Dufresne code's approach of doing "sparse grey dilation" by using the previously-found peak coordinates. In our case we only need to check the pixels outside the rectangular `grey_dilation` footprint. Numba can make that very fast.