@nkeim 2016-05-01
Also, I remember now that @danielballan and I originally decided to do multiple iterations because we were doing peak *merging* — taking the average position of nearby peaks, rather than rejecting the dimmer ones.