@caspervdw 2016-05-06
Apart from the whole `where_close` performance thing (which can be better) I think I understand why you got 546 features with the current method and 351 with the new one.

In fact the current implementation gives unexpected results:

    tp.find.grey_dilation_legacy(im, radius=7)

Would find features upto a separation distance of 7, whereas `locate` expects 15. 
This is because of the following: if you want to find maxima that are spaced 15 from each other, you should do a grey dilation with a circular kernel with a *radius* of 15. It is easy to see in a drawing.

I got it wrong too in the new implementation, so that the KDTree had to do a lot of work.