@caspervdw 2016-05-08
@nkeim, could you check performance again? I made the fixes and I turned off the KDTree filtering in locate, possibly it leads to an improvement even for the dense case.

I am also curious what would happen if you turn in back on, this would mean less features to refine at the expense of an extra `drop_close`.