@nkeim 2016-05-08
Wow. The dense benchmark is now actually *faster* than the v0.3 code — 0.24 s/frame, down from 0.30. The sparse benchmark is still about twice as fast as v0.3, at 0.050 s/frame. Awesome!

I think what @danielballan and I have learned previously is that when using the numba code, characterizing extra features costs almost nothing. So I think it is right to do only the final `drop_close()`. Note, however, that this introduces a new coupling between the peak-finding and refinement steps at a time when we are trying to decouple them. So maybe the default should be to perform KDTree filtering in `grey_dilation`, with an option to turn it off if you promise to do it later?

There is still the disagreement in the results. In the figure, MATLAB is on the left and this new trackpy code is on the right.

<img width="329" alt="unknown" src="https://cloud.githubusercontent.com/assets/2985325/15099318/feb904b8-1506-11e6-8085-b0b5d208d7f1.png">

It looks like I can bring these back into agreement by changing the `percentile` parameter (another parameter for peak-finding). But I first want to make sure I understand what has changed. Am I right that

1. In v0.3, trackpy uses the feature *radius* for grey dilation, so that the overlap between feature ROIs can be as much as one radius; i.e., the minimum separation is one radius. KDTree filtering is applied *after* refinement, and controlled by the `separation` parameter, to set the final separation.
2. In this new code, `diameter` sets the size of the grey dilation kernel. There is still only one KDTree step, applied after refinement.
3. In the Blair-Dufresne code, square kernels are used everywhere, with size specified by a `separation`-like parameter that is independent of `diameter`. Effectively, there is no KDTree step. (I am fairly certain of this based on my reading of the code.)

Is this more or less correct?

Should `separation` set the size of the grey dilation kernel in trackpy? Conceptually, ROI size (for refinement) and neighborhood size (for peak finding and final separation) can be independent, as they are in the Blair-Dufresne code. Being able to set `separation` < `diameter` is useful when it's more important to find all the particles than to know their positions accurately.