@caspervdw 2016-05-08
@nkeim, I think your description is entirely correct (although I have not checked the Blair-Dufresne code in depth)

The most logical choice for the length scale in `grey_dilation` would be `separation`. I cannot come up with any practical use to do it differently. The only thing is that it may lead to issues when combined with `maxsize` filtering (see the failed test `test_minmass_maxsize`, https://travis-ci.org/soft-matter/trackpy/jobs/128360102).

Basically, small features next to large features with size > maxsize will not be detected in the new algorithm, while they would be deteced in the old algorithm. The issue here is a wrong choice of `separation`: when it is much larger than feature `diameter` (as in the test), this leads to the issue.

So I would still go for `separation` and document the slight API change.